--- a/README.md
+++ b/README.md
@@ -49,4 +49,6 @@ mp3: 320kbit/s, 44100 sample rate
 
 - [x] Automatically remove processed files once complete
 
+- [x] Multithreading
+
 - [ ] More?