## Performance Improvements
(with 8 pck files, running on 4.7ghz i7 6700k (4 cores 4 threads))
#### Wav only
```
[Old] 40s
[New] 10s
```
#### Wav and flac
```
[Old] 57s
[New] 14s
```
#### Wav and mp3
```
[Old] 270s
[New] 55s
```
#### Wav, flac and mp3
```
[Old] 277s
[New] 57s
```